+++ src/components/atoms/edit-profile-image/EditProfileImage.tsx
@@ -1,19 +1,12 @@
-import Modal from '@/components/molecules/modal/Modal';
 import { useModal } from '@/store/useModal';
 import Image from 'next/image';
-import Button from '../button/Button';
-import { useImageUploader } from '@/models/uploader/useImageUploader';
+import EditProfileModal from '@/components/molecules/modal/EditProfileModal';
 
 export default function EditProfileImage() {
-  const { uploadedImages, handleUploadImage } = useImageUploader();
-  const { setIsOpen, setIsClose } = useModal();
+  const { setIsOpen } = useModal();
 
   const openModal = (modalKey: string) => {
     setIsOpen(modalKey);
-  };
-
-  const closeModal = () => {
-    setIsClose();
   };
 
   return (
@@ -24,29 +17,7 @@ export default function EditProfileImage() {
       >
         <Image fill src="/profile-image-setting-icon.svg" alt="프로필 사진 수정하기" />
       </button>
-      <Modal modalKey="editProfileImage">
-        <div className="relative flex flex-col items-center justify-center gap-[48px] px-[50px] py-[24px] pb-[72px]">
-          <label className="flex flex-col gap-[32px] font-bold text-[48pxr]">
-            프로필 이미지 변경하기
-            {/* 임시이미지입니다. */}
-            <div className="relative size-[160px] overflow-hidden rounded-full">
-              <Image fill src={uploadedImages[0]} alt="프로필 사진" />
-            </div>
-            <input
-              type="file"
-              accept="image/png, image/jpeg"
-              className="hidden"
-              id="imageUpload"
-              onChange={handleUploadImage}
-            />
-          </label>
-          <div className="absolute bottom-[12px] right-[12px] flex gap-[12px]">
-            <Button size="s" color="white" text="취소" onClick={closeModal} />
-            {/* api 연결 예정*/}
-            <Button size="s" color="black" text="확인" />
-          </div>
-        </div>
-      </Modal>
+      <EditProfileModal />
     </>
   );
 }
